refactor(svg): split elements.ts per element, move isTransparentColor to core, helpers.ts -> utilities.ts
- Split svg/src/elements.ts into path.ts (SVGPath), text.ts (SVGText + SVGTextPath),
  and image.ts (SVGImage); barrel keeps all four exported, tests are barrel-routed.
- Move isTransparentColor into @ripl/core color/utilities.ts (beside setColorAlpha);
  svg consumes it from @ripl/core. It's a generic colour predicate, not svg-specific.
- Rename svg/src/helpers.ts -> utilities.ts for consistent naming.
