+++ packages/svg/src/path.ts
@@ -5,7 +5,6 @@ import type {
 
 import {
     ContextPath,
-    ContextText,
     getThetaPoint,
     normaliseBorderRadius,
     radiansToDegrees,
@@ -13,7 +12,6 @@ import {
 
 import type {
     BorderRadius,
-    TextOptions,
 } from '@ripl/core';
 
 /** SVG-specific path implementation that builds an SVG `d` attribute string from drawing commands. */
@@ -162,85 +160,4 @@ export class SVGPath extends ContextPath implements SVGContextElement {
         this.closePath();
     }
 
-}
-
-/** SVG-specific text element mapping position and content to SVG `<text>` attributes. */
-export class SVGText extends ContextText implements SVGContextElement {
-
-    /** The rendering definition describing this text's SVG `<text>` node. */
-    public definition: SVGContextElementDefinition;
-
-    constructor(options: TextOptions) {
-        super(options);
-
-        // eslint-disable-next-line @typescript-eslint/no-this-alias
-        const _this = this;
-
-        this.definition = {
-            tag: 'text',
-            styles: {
-                fill: 'none',
-            },
-            attributes: {
-                get x() {
-                    return _this.x.toString();
-                },
-                get y() {
-                    return _this.y.toString();
-                },
-            },
-            get textContent() {
-                return _this.pathData ? undefined : _this.content;
-            },
-        };
-    }
-}
-
-/** SVG-specific image element wrapping a `CanvasImageSource` as an SVG `<image>` tag. */
-export class SVGImage implements SVGContextElement {
-
-    /** Unique identifier for this element. */
-    public readonly id: string;
-    /** The rendering definition describing this image's SVG `<image>` node. */
-    public definition: SVGContextElementDefinition;
-
-    constructor(id: string, href: string, x: number, y: number, width: number, height: number) {
-        this.id = id;
-
-        this.definition = {
-            tag: 'image',
-            styles: {},
-            attributes: {
-                href,
-                x: x.toString(),
-                y: y.toString(),
-                width: width.toString(),
-                height: height.toString(),
-                preserveAspectRatio: 'none',
-            },
-        };
-    }
-}
-
-/** SVG `<textPath>` element for rendering text along a path defined in `<defs>`. */
-export class SVGTextPath implements SVGContextElement {
-
-    /** Unique identifier for this element, derived from the owning text element's id. */
-    public readonly id: string;
-    /** The rendering definition describing this element's SVG `<textPath>` node. */
-    public definition: SVGContextElementDefinition;
-
-    constructor(textId: string, pathId: string, content: string, startOffset?: number) {
-        this.id = `${textId}:textpath`;
-
-        this.definition = {
-            tag: 'textPath',
-            styles: {},
-            attributes: {
-                href: `#${pathId}`,
-                ...(startOffset !== undefined ? { startOffset: `${startOffset * 100}%` } : {}),
-            },
-            textContent: content,
-        };
-    }
 }
